fix: split client chess hook to resolve build error
